packetizer: hxxx: remove bogus cast

+++ modules/packetizer/hxxx_nal.h
@@ -284,7 +284,7 @@ static inline bool hxxx_iterate_next( hxxx_iterator_ctx_t *p_ctx, const uint8_t 
     for( uint8_t i=0; i < p_ctx->i_nal_length_size ; i++ )
         i_nal_size = (i_nal_size << 8) | *p_ctx->p_head++;
 
-    if( (ptrdiff_t) i_nal_size > p_ctx->p_tail - p_ctx->p_head )
+    if( i_nal_size > p_ctx->p_tail - p_ctx->p_head )
         return false;
 
     *pp_start = p_ctx->p_head;
